Update Export and Printing Instructions
Update the instructions to print or export a plain copy of the resume from my
new Macbook.

Add instructions to print or export a 'pretty' copy of the resume (with some
minimal formatting).  Currently, the 'pretty' copy only bolds the
section titles but otherwise keeps the same Markdown plaintext format.

Change the max characters per line from 98 to 97 (was previously
incorrect).